// data/interfaces/projects/projects_interfaces.ts
export interface Projects {
  id: String
  name: String
  phone: String
  title_project: String
  type_project: String
  description: String
  reference: String
  estimation: String
  range_min_fee: String
  range_max_fee: String
  srs: String
  status: String
  inserted_at: Date
  updated_at: Date
}